% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{as_biadjacency.list}
\alias{as_biadjacency.list}
\title{Given a list of vectors, this function returns a biadjacency matrix
indicating whether a given element (row) is present in a given vector
(column)}
\usage{
\method{as_biadjacency}{list}(x, universe = NULL, ...)
}
\arguments{
\item{x}{A list of vectors; each entry should be a subset
of \code{universe}.}

\item{universe}{A vector. Incidence mappings will only be
returned for entries that are present in this vector. If missing/NULL, the
union of the genesets is taken to be the universe.}

\item{...}{Further arguments (unused at present).}
}
\description{
Given a list of vectors, this function returns a biadjacency matrix
indicating whether a given element (row) is present in a given vector
(column)
}
