% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_set_functions.R
\name{as_biadjacency}
\alias{as_biadjacency}
\title{Generic function `as_biadjacency` which returns a biadjacency / incidence
matrix}
\usage{
as_biadjacency(x, ...)
}
\arguments{
\item{x}{An object to turn into a biadjacency Matrix}

\item{...}{Further arguments.}
}
\description{
Generic function `as_biadjacency` which returns a biadjacency / incidence
matrix
}
