% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{merge.matrix}
\alias{merge.matrix}
\title{Generic function `merge` for merging two matrices by row-names}
\usage{
\method{merge}{matrix}(x, y, ...)
}
\arguments{
\item{x, y}{Matrices.}

\item{...}{Arguments for passing through to base::merge
(the data.frame method). Do not specify a `by` argument.}
}
\description{
A matrix is returned.
}
